+++ integration_tests/kafka/test_simple.py
@@ -5,7 +5,11 @@ import pathlib
 
 import pathway as pw
 from pathway.internals.parse_graph import G
-from pathway.tests.utils import expect_csv_checker, wait_result_with_checker
+from pathway.tests.utils import (
+    FileLinesNumberChecker,
+    expect_csv_checker,
+    wait_result_with_checker,
+)
 
 from .utils import KafkaTestContext
 
@@ -16,7 +20,7 @@ def test_kafka_raw(tmp_path: pathlib.Path, kafka_context: KafkaTestContext):
     table = pw.io.kafka.read(
         rdkafka_settings=kafka_context.default_rdkafka_settings(),
         topic=kafka_context.input_topic,
-        format="raw",
+        format="plaintext",
         autocommit_duration_ms=100,
     )
 
@@ -110,28 +114,17 @@ def test_kafka_csv(tmp_path: pathlib.Path, kafka_context: KafkaTestContext):
     )
 
 
-def test_kafka_simple_wrapper(tmp_path: pathlib.Path, kafka_context: KafkaTestContext):
+def test_kafka_simple_wrapper_bytes_io(
+    tmp_path: pathlib.Path, kafka_context: KafkaTestContext
+):
     kafka_context.fill(["foo", "bar"])
 
     table = pw.io.kafka.simple_read(
         "kafka:9092",
         kafka_context.input_topic,
     )
-    pw.io.csv.write(table, str(tmp_path / "output.csv"))
-
-    wait_result_with_checker(
-        expect_csv_checker(
-            """
-            data
-            foo
-            bar
-            """,
-            tmp_path / "output.csv",
-            usecols=["data"],
-            index_col=["data"],
-        ),
-        10,
-    )
+    pw.io.jsonlines.write(table, str(tmp_path / "output.jsonl"))
+    wait_result_with_checker(FileLinesNumberChecker(tmp_path / "output.jsonl", 2), 10)
 
     # check that reread will have all these messages again
     G.clear()
@@ -139,21 +132,48 @@ def test_kafka_simple_wrapper(tmp_path: pathlib.Path, kafka_context: KafkaTestCo
         "kafka:9092",
         kafka_context.input_topic,
     )
-    pw.io.csv.write(table, str(tmp_path / "output.csv"))
+    pw.io.jsonlines.write(table, str(tmp_path / "output.jsonl"))
+    wait_result_with_checker(FileLinesNumberChecker(tmp_path / "output.jsonl", 2), 10)
 
-    wait_result_with_checker(
-        expect_csv_checker(
-            """
-            data
-            foo
-            bar
-            """,
-            tmp_path / "output.csv",
-            usecols=["data"],
-            index_col=["data"],
-        ),
-        10,
+    # Check output type, bytes should be rendered as an array
+    with open(tmp_path / "output.jsonl", "r") as f:
+        for row in f:
+            row_parsed = json.loads(row)
+            assert isinstance(row_parsed["data"], list)
+            assert row_parsed["data"] == list(b"foo") or row_parsed["data"] == list(
+                b"bar"
+            )
+
+
+def test_kafka_simple_wrapper_plaintext_io(
+    tmp_path: pathlib.Path, kafka_context: KafkaTestContext
+):
+    kafka_context.fill(["foo", "bar"])
+
+    table = pw.io.kafka.simple_read(
+        "kafka:9092",
+        kafka_context.input_topic,
+        format="plaintext",
     )
+    pw.io.jsonlines.write(table, str(tmp_path / "output.jsonl"))
+    wait_result_with_checker(FileLinesNumberChecker(tmp_path / "output.jsonl", 2), 10)
+
+    # check that reread will have all these messages again
+    G.clear()
+    table = pw.io.kafka.simple_read(
+        "kafka:9092",
+        kafka_context.input_topic,
+        format="plaintext",
+    )
+    pw.io.jsonlines.write(table, str(tmp_path / "output.jsonl"))
+    wait_result_with_checker(FileLinesNumberChecker(tmp_path / "output.jsonl", 2), 10)
+
+    # Check output type, parsed plaintext should be a string
+    with open(tmp_path / "output.jsonl", "r") as f:
+        for row in f:
+            row_parsed = json.loads(row)
+            assert isinstance(row_parsed["data"], str)
+            assert row_parsed["data"] == "foo" or row_parsed["data"] == "bar"
 
 
 def test_kafka_output(tmp_path: pathlib.Path, kafka_context: KafkaTestContext):
@@ -169,6 +189,56 @@ def test_kafka_output(tmp_path: pathlib.Path, kafka_context: KafkaTestContext):
         table,
         rdkafka_settings=kafka_context.default_rdkafka_settings(),
         topic_name=kafka_context.output_topic,
+    )
+    pw.run()
+
+    output_topic_contents = kafka_context.read_output_topic()
+    assert len(output_topic_contents) == 2, output_topic_contents
+
+
+def test_kafka_raw_bytes_output(
+    tmp_path: pathlib.Path, kafka_context: KafkaTestContext
+):
+    input_path = tmp_path / "input"
+    input_path.mkdir()
+    (input_path / "foo").write_text("foo")
+    (input_path / "bar").write_text("bar")
+
+    table = pw.io.fs.read(
+        input_path,
+        mode="static",
+        format="binary",
+    )
+    pw.io.kafka.write(
+        table,
+        rdkafka_settings=kafka_context.default_rdkafka_settings(),
+        topic_name=kafka_context.output_topic,
+        format="raw",
+    )
+    pw.run()
+
+    output_topic_contents = kafka_context.read_output_topic()
+    assert len(output_topic_contents) == 2, output_topic_contents
+
+
+def test_kafka_plaintext_output(
+    tmp_path: pathlib.Path, kafka_context: KafkaTestContext
+):
+    input_path = tmp_path / "input"
+    input_path.mkdir()
+    (input_path / "foo").write_text("foo")
+    (input_path / "bar").write_text("bar")
+
+    table = pw.io.fs.read(
+        input_path,
+        mode="static",
+        format="plaintext",
+    )
+    pw.io.kafka.write(
+        table,
+        rdkafka_settings=kafka_context.default_rdkafka_settings(),
+        topic_name=kafka_context.output_topic,
+        format="plaintext",
     )
     pw.run()